+++ node/CHANGELOG.md
@@ -12,6 +12,34 @@ All notable changes to this project will be documented in this file.  The format
 
 ## [Unreleased]
 
+
+
+## 1.4.13
+
+### Changed
+* Update `casper-execution-engine`.
+
+
+
+## 1.4.8
+
+### Changed
+* Update `casper-execution-engine`.
+
+
+
+## 1.4.7
+
+### Changed
+* Update `casper-execution-engine` and three `openssl` crates to latest versions.
+
+
+
+## 1.4.6
+
+### Changed
+* Update dependencies to make use of scratch global state in the contract runtime.
+
 ### Added
 * Introduce fast-syncing to join the network, avoiding the need to execute every block to catch up.
 * Add config sections for new components to support fast-sync: `[block_accumulator]`, `[block_synchronizer]`, `[deploy_buffer]` and `[upgrade_watcher]`.